requirements.txt
requests>=2.31.0
google-auth>=2.23.0
google-api-python-client>=2.100.0
# schedule>=1.2.0
# pytz>=2024.1
# beautifulsoup4>=4.12.0
